+++ lib/index.js
@@ -183,7 +183,7 @@ module.exports = {
     'script-setup-uses-vars': require('./rules/script-setup-uses-vars'),
     'singleline-html-element-content-newline': require('./rules/singleline-html-element-content-newline'),
     'sort-keys': require('./rules/sort-keys'),
-    'space-between-siblings': require('./rules/space-between-siblings'),
+    'padding-line-between-tags': require('./rules/padding-line-between-tags'),
     'space-in-parens': require('./rules/space-in-parens'),
     'space-infix-ops': require('./rules/space-infix-ops'),
     'space-unary-ops': require('./rules/space-unary-ops'),
